refactor: 홈 API를 독립 실행 컨텍스트로 분리
홈 화면 조회는 performance aggregate API가 아니라 promotion/schedule/performance를 조합하는 API-facing context라서 home 패키지로 이동했다. URL, query parameter, JSON response, success message는 기존 계약을 유지한다.

Constraint: /api/main endpoint and client-facing JSON contract must remain unchanged

Rejected: keep Home under performance | performance package would continue to own home composition concerns

Confidence: high

Scope-risk: narrow

Directive: Place new home-facing controller/facade/application/response code under com.beat.apis.home unless it is a shared API enum or lower domain query dependency.

Tested: ./gradlew :apis:compileJava :apis:compileKotlin --no-daemon; ./gradlew :apis:test --tests com.beat.apis.ApisDtoJsonContractTest --tests com.beat.apis.boundary.DomainApplicationCodeBoundarySnapshotTest --tests com.beat.apis.ApisArchitectureGuardTest --no-daemon; ./gradlew check --no-daemon; git diff --check

Not-tested: live HTTP request against a running server
