+++ apis/src/main/java/com/beat/apis/home/application/dto/HomePerformanceDetail.java
@@ -1,4 +1,4 @@
-package com.beat.apis.performance.application.dto.home;
+package com.beat.apis.home.application.dto;
 
 public record HomePerformanceDetail(
 	Long performanceId,
