+++ apis/src/main/java/com/beat/apis/home/application/dto/HomePromotionDetail.java
@@ -1,7 +1,13 @@
-package com.beat.apis.performance.application.dto.home;
+package com.beat.apis.home.application.dto;
 
-public record HomePromotionDetail(Long promotionId, String promotionPhoto, Long performanceId, String redirectUrl,
-								  boolean isExternal, String carouselNumber) {
+public record HomePromotionDetail(
+	Long promotionId,
+	String promotionPhoto,
+	Long performanceId,
+	String redirectUrl,
+	boolean isExternal,
+	String carouselNumber
+) {
 
 	public static HomePromotionDetail of(Long promotionId, String promotionPhoto, Long performanceId, String redirectUrl,
 		boolean isExternal, String carouselNumber) {
