+++ apis/src/test/java/com/beat/apis/ApisDtoJsonContractTest.java
@@ -12,10 +12,10 @@ import org.junit.jupiter.api.Test;
 import com.beat.apis.booking.application.dto.BookingRefundRequest;
 import com.beat.apis.booking.application.dto.GuestBookingRequest;
 import com.beat.apis.member.application.dto.request.MemberLoginRequest;
-import com.beat.apis.performance.application.dto.home.HomeFindAllResponse;
-import com.beat.apis.performance.application.dto.home.HomeFindRequest;
-import com.beat.apis.performance.application.dto.home.HomePerformanceDetail;
-import com.beat.apis.performance.application.dto.home.HomePromotionDetail;
+import com.beat.apis.home.application.dto.HomeFindAllResponse;
+import com.beat.apis.home.application.dto.HomeFindRequest;
+import com.beat.apis.home.application.dto.HomePerformanceDetail;
+import com.beat.apis.home.application.dto.HomePromotionDetail;
 import com.beat.apis.booking.application.dto.BookingStatusType;
 import com.beat.apis.member.application.dto.request.SocialTypeRequest;
 import com.beat.apis.performance.application.dto.BankNameType;
